My first try to launch Spring application
