+++ src/main/java/controller/Launcher.java
@@ -1,0 +1,11 @@
+package controller;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+@SpringBootApplication
+public class Launcher {
+    public static void main(String[] args) {
+        SpringApplication.run(Launcher.class, args);
+    }
+}
