+++ src/main/java/controller/RecepiesService.java
@@ -1,3 +1,7 @@
+package controller;
+
+import service.Recepie;
+
 import java.util.List;
 
 public interface RecepiesService {
