+++ src/main/java/service/ConsoleInput.java
@@ -1,3 +1,5 @@
+package service;
+
 import java.util.Scanner;
 
 public class ConsoleInput implements Input {
