+++ src/main/java/service/JSONRecepieConverter.java
@@ -1,3 +1,5 @@
+package service;
+
 import com.google.gson.Gson;
 import com.google.gson.GsonBuilder;
