+++ src/main/java/service/ReceiptsJSONFileStorage.java
@@ -1,3 +1,5 @@
+package service;
+
 import java.io.*;
 import java.nio.file.Files;
 import java.nio.file.Path;
