+++ src/main/java/service/RecepiesStorage.java
@@ -1,3 +1,5 @@
+package service;
+
 import java.util.List;
 
 public interface RecepiesStorage {
