+++ src/main/java/service/Saver.java
@@ -1,3 +1,5 @@
+package service;
+
 public interface Saver {
     void saveData(String data);
 }
