+++ src/test/java/JSONRecepieConverterTest.java
@@ -1,5 +1,6 @@
 import org.junit.Assert;
 import org.junit.Test;
+import service.*;
 
 import java.util.ArrayList;
 import java.util.List;
